@@ -29,7 +29,8 @@
     "css-modules-require-hook": "4.0.6",
     "deep-extend": "^0.5.0",
     "empty": "^0.10.1",
-    "enzyme": "^2.4.1",
+    "enzyme": "^3.3.0",
+    "enzyme-adapter-react-16": "^1.1.1",
     "es6-promise": "^4.0.5",
     "es6-promisify": "^5.0.0",
     "eslint": "^3.11.1",
@@ -77,17 +78,18 @@
     "postcss-loader": "^2.0.9",
     "pre-git": "^3.10.0",
     "prettier": "^1.8.2",
+    "prop-types": "^15.5.7",
     "proxyquire": "^1.8.0",
     "ramda": "^0.22.1",
     "raw-loader": "^0.5.1",
-    "react": "^15.3.2",
-    "react-addons-test-utils": "^15.3.2",
-    "react-dom": "^15.3.1",
+    "react": "^16.0.0",
+    "react-dom": "^16.0.0",
+    "react-addons-test-utils": "^16.0.0",
     "react-ga": "^2.2.0",
-    "react-helmet": "^4.0.0",
+    "react-helmet": "^5.2.0",
     "react-modal": "^3.1.6",
     "react-motion": "^0.4.7",
-    "react-swipeable-views": "^0.10.1",
+    "react-swipeable-views": "^0.12.13",
     "reset-css": "^2.2.0",
     "sanitize-html": "^1.16.3",
     "serverless": "1.0.0-rc.1",
