@@ -1,9 +1,10 @@
 import React, { Component } from 'react';
+import PropTypes from 'prop-types';
 import classnames from 'classnames/bind';
 import styles from '../style.css';
 
 const cx = classnames.bind(styles);
-const { func, string, bool } = React.PropTypes;
+const { func, string, bool } = PropTypes;
 
 const AfterSignUpForm = ({ errorMessage, handleInputChange, submitting, handleSubmit }) => (
   <div>
