@@ -1,8 +1,11 @@
 // @flow
 import React from 'react';
 import { expect } from 'chai';
-import { shallow } from 'enzyme';
+import Enzyme, { shallow } from 'enzyme';
+import Adapter from 'enzyme-adapter-react-16';
 import Avatar from '.';
+
+Enzyme.configure({ adapter: new Adapter() });
 
 describe('<Avatar />', () => {
   it('renders image child', () => {
