@@ -1,10 +1,13 @@
 import React from 'react';
-import { shallow } from 'enzyme';
+import Enzyme, { mount } from 'enzyme';
+import Adapter from 'enzyme-adapter-react-16';
 import AfterSignUp from '.';
+
+Enzyme.configure({ adapter: new Adapter() });
 
 describe('AfterSignUp', () => {
   it('sets the initial state correctly', () => {
-    const wrapper = shallow(<AfterSignUp />);
+    const wrapper = mount(<AfterSignUp />);
     expect(wrapper.state()).to.deep.equal({
       name: '',
       company: '',
@@ -14,7 +17,7 @@ describe('AfterSignUp', () => {
   });
 
   it('handles handleInputChange and sets the state correctly', () => {
-    const wrapper = shallow(<AfterSignUp />);
+    const wrapper = mount(<AfterSignUp />);
     const event = {
       target: {
         name: 'exampleName',
@@ -32,7 +35,7 @@ describe('AfterSignUp', () => {
   });
 
   it('handles componentWillReceiveProps and sets the state correctly', () => {
-    const wrapper = shallow(<AfterSignUp />);
+    const wrapper = mount(<AfterSignUp />);
     wrapper.instance().componentWillReceiveProps();
     expect(wrapper.state()).to.deep.equal({
       name: '',
@@ -56,7 +59,7 @@ describe('AfterSignUp', () => {
           },
         });
       });
-    const wrapper = shallow(<AfterSignUp onSubmit={submitFormFunction} />);
+    const wrapper = mount(<AfterSignUp onSubmit={submitFormFunction} />);
     wrapper.instance().handleSubmit();
     expect(wrapper.state()).to.deep.equal({
       submitting: true,
