@@ -1,4 +1,5 @@
-import React, { PropTypes } from 'react';
+import React from 'react';
+import PropTypes from 'prop-types';
 import classnames from 'classnames/bind';
 import Link from '../../../components/link';
 import styles from './style.css';
